Fixes some issues posed by NAs
The changes to the CleaningCensusInfo function in the last commit meant
that stems that have not died yet or been snapped yet had NA values for
IsSnapped and CensusStemDied/ CensusNoDead. This causes problems
downstream as these trees are coded NA for Alive rather than 1. I have
fixed this by setting trees with NA for CensusStemDied and CensusNoDead
(i.e. trees that have not yet died) to 9999 (an high number larger than
any census number), and trees with NA for IsSnapped (i.e. trees that
have not yet snapped) to 0. This appears to fix downstream problems.